refactor(metrics): move LoggingMetricsReporter to extensions/metrics-reports/base
Now that the IcebergMetricsReporter SPI lives in its own module,
extensions/metrics-reports/base no longer needs runtime/service for
anything, and the default logging implementation can live alongside
its peers (NoOpMetricsReporter, and PersistingMetricsReporter in the
relational-jdbc extension) instead of in the Quarkus runtime module.

Adds guava to the module's own dependencies, for the @VisibleForTesting
that LoggingMetricsReporter carries on its test constructor. It reaches
this module only through a transitive implementation configuration
today, which doesn't put it on a consumer's compile classpath. Same
pattern as extensions/auth/opa and extensions/semantic-models.
